refactor: 타입 파일로 분리
여러군데서 사용되는 타입을 파일로 분리
