+++ src/contexts/HabitContextProvider.tsx
@@ -6,26 +6,14 @@ import React, {
   useMemo,
   useState,
 } from 'react';
-import { usePeriodValue, useToday } from './PeriodProvider';
-
-export type Day = '월' | '화' | '수' | '목' | '금' | '토' | '일';
-export interface HabitType {
-  id: number;
-  name: string;
-  description: string;
-  days: Day[];
-  checks: string[]; // yyyy-mm-dd string
-}
-
-export type HabitCreateDataType = Pick<
+import {
+  HabitCreateDataType,
+  HabitCheckChangeDataType,
   HabitType,
-  'name' | 'description' | 'days'
->;
-export type HabitCheckChangeDataType = {
-  id: HabitType['id'];
-  date: string;
-  value: boolean;
-};
+  HabitService,
+  Day,
+} from '../service/types';
+import { usePeriodValue, useToday } from './PeriodProvider';
 
 interface HabitsActionContextValue {
   addHabit: (habit: HabitCreateDataType) => void;
@@ -51,13 +39,6 @@ const HabitsTodayContext = createContext<HabitType[] | null>(null);
 const HabitsActionContext = createContext<HabitsActionContextValue | null>(
   null,
 );
-
-export interface HabitService {
-  getHabitsByPeriod: (startDate: string, endDate: string) => HabitType[];
-  getHabitsByDay: (day: Day) => HabitType[];
-  addHabit: (habit: HabitCreateDataType) => void;
-  changeHabitCheck: ({ id, date, value }: HabitCheckChangeDataType) => void;
-}
 
 const HabitContextProvider = ({
   children,
